update: fix path db config

+++ P10/php10H.php
@@ -3,14 +3,21 @@
 <!-- 2KS1 -->
 
 <?php
-require_once "../db.php";
+session_start();
+require_once "db.php";
+
+if (!isset($_SESSION['username'])) {
+	header("Location: php10A.php");
+	exit();
+}
 
 try {
 	$stmt = $pdo->prepare("DELETE FROM meetings WHERE slot = :slot");
-	$stmt->execute(array(':slot' => $_GET['slot']));
+	$stmt->bindParam(':slot', $_GET['slot'], PDO::PARAM_INT);
+	$stmt->execute();
 
 	if ($stmt->rowCount() == 1) {
-		return header("Location: php09F.php");
+		return header("Location: php10F.php");
 	} else {
 		echo "Gagal menghapus data<br>";
 	}
